requirements.txt
google-genai
polars
Pillow
PyMuPDF
matplotlib
ratelimit
streamlit
streamlit-option-menu
